moved stuff around, fixed missing parts
