+++ Cargo.toml
@@ -49,6 +49,7 @@ members = [
   "examples/node_refs",
   "examples/npm_and_rest",
   "examples/pub_sub",
+  "examples/store",
   "examples/textarea",
   "examples/timer",
   "examples/todomvc",
