--- examples/stateful-camera/bootstrap.js
@@ -1,3 +1,0 @@
-import("./pkg").then(module => {
-  module.run_app();
-});
