--- examples/stateful-camera/src/lib.rs
@@ -1,16 +1,0 @@
-mod app;
-mod agents;
-
-use wasm_bindgen::prelude::*;
-
-#[cfg(feature = "wee_alloc")]
-#[global_allocator]
-static ALLOC: wee_alloc::WeeAlloc = wee_alloc::WeeAlloc::INIT;
-
-#[wasm_bindgen]
-pub fn run_app() -> Result<(), JsValue> {
-    yew::initialize();
-    yew::start_app::<app::App>();
-
-    Ok(())
-}
