+++ examples/store/src/agents/mod.rs
