+++ examples/store/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    yew::start_app::<store::App>();
+}
