+++ examples/store/static/index.html
