+++ yew/src/agent/local/mod.rs
@@ -1,9 +1,7 @@
 mod context;
 mod job;
-mod store;
 
 use super::*;
 
 pub use context::Context;
 pub use job::Job;
-pub use store::{ReadOnly, Store, StoreWrapper, Bridgeable};
