+++ yew/src/agent/mod.rs
@@ -7,7 +7,7 @@ mod worker;
 
 pub use link::AgentLink;
 pub(crate) use link::*;
-pub use local::{Context, Job, ReadOnly, Store, StoreWrapper, Bridgeable};
+pub use local::{Context, Job};
 pub(crate) use pool::*;
 pub use pool::{Dispatched, Dispatcher};
 pub use worker::{Private, Public, Threaded};
